test(web): unmount the viewer in integration boot tests
The two integration tests that boot the whole viewer by importing
src/viewer/main.ts never unmounted it. Clearing document.body detaches the
DOM but leaves the Preact tree live with an effect flush still queued:
Preact 10.29.8 defers flushes via requestAnimationFrame with a
setTimeout(..., 35) fallback, and vitest jsdom has no rAF, so the fallback
always runs. When the file finishes first, vitest tears down the
environment and the flush then executes effect bodies with no window on
globalThis, raising an uncaught ReferenceError that fails CI even though
every test passes.

Add tests/helpers/viewer-boot.ts with bootViewer/teardownViewer. Teardown
does render(null, root) -- so hook cleanups actually unwind -- then drains
50ms while jsdom is alive. bootViewer tears down first, so repeated boots
inside one test no longer orphan a tree.

The helper takes render from a dynamic import(preact) after main.ts is
loaded: vi.resetModules() gives main.ts a fresh Preact instance per boot, so
a static top-level import would not own the mounted tree.

Both files also duplicated their stub helpers verbatim; those now live in
the shared helper. Adds one test asserting the container is gone after
teardown, so a regression of the harness fails deterministically instead of
as CI flake.

Test-only; no production code changes.
